Modal Regions inherit from a common base
